Use safe wrapping
The wrap code is now fully separated and only unsafely loads traceback.
Now if any exception is thrown it will catch it. Hopefully it will be a
`HTTPResponse` which will be rendered correctly. If not it falls back to
full server error and simple text output with traceback.

`wrap.py` SHOULD NOT BE TOUCHED!! It works and will always work! You do
not want errors that just don't appear.