Implement file uploading
Sadly `python-cgi` is discontinued but it is possible to use
`email.message.Message` and more importantly `multipart-python` to get
the job done.

This requires a fair amount of checks and gathering of data to get a
clean parsing of the parameters for both expected HTTP calls and the
different POST content types. Unfortunately, [`multipart-python` does not
currently support decoding file mimetypes even though this is standardly
included in `multipart/form-data`](https://github.com/Kludex/python-multipart/issues/207)
. I opted to assume it was always an image as for now it will be. Using
the file extension could also be an indicator.