Implement JWT and cookies
Implement JSON web tokens [rfc7518](https://www.rfc-editor.org/rfc/rfc7518) to store an authorised newsletter login. The user is authenticated once on unlock and the relevant newsletter data (id, name, folder, etc) are stored in a JWT which is saved as a cookie. Now submissions do not need to secretly hold the passcode (requiring unnecessary authorisations) and paves the way for history view.

This also required the implementation of CGI cookie parsing and setting,
as well as a foray into HTTP statuses to redirect the user. Additional
work to timeout tokens should be taken to prevent tokens unlocking old
issues in the wrong mode.